fix:admin_crud_story
