+++ app/Http/Controllers/Api/Admin/AdminGenreController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Controllers\Api\Admin\GenreManagement;
+namespace App\Http\Controllers\Api\Admin;
 
 use App\Http\Controllers\Controller;
 use App\Http\Resources\Api\Genre\GenreResource;
