+++ app/Interfaces/Author/AuthorRepositoryInterface.php
@@ -6,4 +6,5 @@ use App\Interfaces\CrudRepositoryInterface;
 
 interface AuthorRepositoryInterface extends CrudRepositoryInterface
 {
+    public function getAuthorByData($colum, $data, $compare = '=');
 }
