+++ app/Interfaces/Book/BookRepositoryInterface.php
@@ -6,5 +6,5 @@ use App\Interfaces\CrudRepositoryInterface;
 
 interface BookRepositoryInterface extends CrudRepositoryInterface
 {
-    //
+    public function getMyBooks(int $userId);
 }
