+++ app/Interfaces/User/UserRepositoryInterface.php
@@ -9,8 +9,4 @@ interface UserRepositoryInterface extends CrudRepositoryInterface
     public function findByEmail($email);
 
     public function vertifyEmail($token);
-
-    public function getListAuthors($id);
-
-    public function getMyBooks($id);
 }
