+++ app/Repositories/Author/AuthorRepository.php
@@ -12,4 +12,9 @@ class AuthorRepository extends BaseRepository implements AuthorRepositoryInterfa
     {
         $this->model = $author;
     }
+
+    public function getAuthorByData($colum, $data, $compare = '=')
+    {
+        return $this->model->where($colum, $compare, $data);
+    }
 }
